Remove UserConfiguration class and related EF Core setup
The UserConfiguration class and its implementation have been removed. This includes the removal of the using directives for Microsoft.EntityFrameworkCore and Microsoft.EntityFrameworkCore.Metadata.Builders, as well as the removal of the UserConfiguration class that implemented the IEntityTypeConfiguration<User> interface. The Configure method, which set up the primary key for the User entity, has also been removed.