Remove User class and related using directives
The User class, which was part of the DataAccess.Entities namespace and implemented the IEntity interface, has been completely removed. This includes the removal of its properties (Id, Name, and Email) and the Column attribute associated with the Id property. Additionally, the using directives for DataAccess.Interfaces and System.ComponentModel.DataAnnotations.Schema have also been removed as they were only needed for the User class.